test: add tests for jwt cookie
